refactor(net_sim_test): Add type to the lossy networks generalized C&C strategy.

+++ theboss/simulation_strategies/lossy_networks_generalized_cliffords_simulation_strategy.py
@@ -15,7 +15,7 @@ class LossyNetworksGeneralizedCliffordsSimulationStrategy(SimulationStrategyInte
         self._helper_strategy = GeneralizedCliffordsSimulationStrategy(bs_permanent_calculator)
 
     def simulate(self, input_state: ndarray, samples_number: int = 1) -> [ndarray]:
-        expansion_zeros = zeros_like(input_state)
+        expansion_zeros = zeros_like(input_state, dtype=int)
         expanded_state = vstack([input_state, expansion_zeros])
         expanded_state = expanded_state.reshape(2 * len(input_state), )
         expanded_samples = self._helper_strategy.simulate(expanded_state, samples_number)
